added first tests for unbalanced calculations
